feat: changed decimal to doubles; implement PairsCointegrationStrategy, implemented a IGpuService interface for allow parallel processing when doing many calcs at one.
Implemented an IBaseStrategy class to handle the different request types.

+++ Temperance.Data/Data/Repositories/Trade/Implementations/BacktestRepository.cs
@@ -18,7 +18,7 @@ namespace Temperance.Data.Data.Repositories.Trade.Implementations
             _logger = logger;
         }
 
-        public async Task UpdateBacktestPerformanceMetrics(Guid runId, BacktestResult result, decimal initialCapital)
+        public async Task UpdateBacktestPerformanceMetrics(Guid runId, BacktestResult result, double initialCapital)
         {
             const string query = @"
                 UPDATE [TradingBotDb].[Constellations].[BacktestRuns]
