+++ Temperance.Services/BackTesting/Interfaces/IPerformanceCalculator.cs
@@ -6,7 +6,7 @@ namespace Temperance.Services.BackTesting.Interfaces
 {
     public interface IPerformanceCalculator
     {
-        Task CalculatePerformanceMetrics(BacktestResult result, decimal initialCapital);
+        Task CalculatePerformanceMetrics(BacktestResult result, double initialCapital);
 
         KellyMetrics CalculateKellyMetrics(IReadOnlyList<TradeSummary> trades);
     }
